+++ apps/basecamp-tokenstudio/flake.nix
@@ -4,8 +4,9 @@
   inputs = {
     logos-module-builder.url = "github:logos-co/logos-module-builder";
     nix-bundle-lgx.url = "github:logos-co/nix-bundle-lgx";
-    logos_execution_zone.url = "github:logos-blockchain/logos-execution-zone-module";
+    lez_core.url = "github:logos-blockchain/logos-execution-zone-module";
     chat_module.url = "github:logos-co/logos-chat-module";
+    delivery_module.follows = "chat_module/logos-delivery-module";
   };
 
   outputs = inputs@{ logos-module-builder, ... }: